requirements.txt
streamlit>=1.31
python-dotenv>=1.0
openai>=1.40.0
tenacity>=8.5.0
pypdf>=4.2.0
python-docx>=1.1.2
pydantic>=2.7.0
requests>=2.32.0
